Rename sample
