+++ samples/NumberParser/NumberParser.sln
@@ -3,16 +3,16 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.645
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "ConverterTest", "ConverterTest\ConverterTest.nfproj", "{2A5794EF-77E7-41A9-8BC6-1077DB0499FE}"
+Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "NumberParserTest", "NumberParserTest\NumberParserTest.nfproj", "{2A5794EF-77E7-41A9-8BC6-1077DB0499FE}"
 EndProject
-Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "ConverterTestShared", "ConverterTestShared\ConverterTestShared.shproj", "{E98BCE69-4C45-4D98-B38E-C3ED8C04EFFF}"
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "NumberParserTestShared", "NumberParserTestShared\NumberParserTestShared.shproj", "{E98BCE69-4C45-4D98-B38E-C3ED8C04EFFF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DesktopConverterTest", "DesktopConverterTest\DesktopConverterTest.csproj", "{B7FEF664-44B1-4E33-8F16-8D5B6DA7A11D}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DesktopNumberParserTest", "DesktopNumberParserTest\DesktopNumberParserTest.csproj", "{B7FEF664-44B1-4E33-8F16-8D5B6DA7A11D}"
 EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
-		ConverterTestShared\ConverterTestShared.projitems*{b7fef664-44b1-4e33-8f16-8d5b6da7a11d}*SharedItemsImports = 4
-		ConverterTestShared\ConverterTestShared.projitems*{e98bce69-4c45-4d98-b38e-c3ed8c04efff}*SharedItemsImports = 13
+		NumberParserTestShared\NumberParserTestShared.projitems*{b7fef664-44b1-4e33-8f16-8d5b6da7a11d}*SharedItemsImports = 4
+		NumberParserTestShared\NumberParserTestShared.projitems*{e98bce69-4c45-4d98-b38e-c3ed8c04efff}*SharedItemsImports = 13
 	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
