+++ samples/NumberParser/NumberParserTest/NumberParserTest.nfproj
@@ -12,8 +12,8 @@
     <OutputType>Exe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <FileAlignment>512</FileAlignment>
-    <RootNamespace>ConverterTest</RootNamespace>
-    <AssemblyName>ConverterTest</AssemblyName>
+    <RootNamespace>NumberParserTest</RootNamespace>
+    <AssemblyName>NumberParserTest</AssemblyName>
     <TargetFrameworkVersion>v1.0</TargetFrameworkVersion>
   </PropertyGroup>
   <Import Project="$(NanoFrameworkProjectSystemPath)NFProjectSystem.props" Condition="Exists('$(NanoFrameworkProjectSystemPath)NFProjectSystem.props')" />
@@ -31,7 +31,7 @@
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
-  <Import Project="..\ConverterTestShared\ConverterTestShared.projitems" Label="Shared" />
+  <Import Project="..\NumberParserTestShared\NumberParserTestShared.projitems" Label="Shared" />
   <Import Project="$(NanoFrameworkProjectSystemPath)NFProjectSystem.CSharp.targets" Condition="Exists('$(NanoFrameworkProjectSystemPath)NFProjectSystem.CSharp.targets')" />
   <ProjectExtensions>
     <ProjectCapabilities>
