+++ samples/NumberParser/NumberParserTest/Properties/AssemblyInfo.cs
