+++ samples/NumberParser/NumberParserTest/packages.config
