+++ samples/NumberParser/NumberParserTestShared/NumberParserTestShared.projitems
@@ -6,7 +6,7 @@
     <SharedGUID>e98bce69-4c45-4d98-b38e-c3ed8c04efff</SharedGUID>
   </PropertyGroup>
   <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>ConverterTestShared</Import_RootNamespace>
+    <Import_RootNamespace>NumberParserTestShared</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)TestInt16.cs" />
