+++ samples/NumberParser/NumberParserTestShared/TestInt32.cs
@@ -1,23 +1,23 @@
 ﻿using System;
 
-namespace ConverterTestShared
+namespace NumberParserTestShared
 {
-    public class TestUInt32
+    public class TestInt32
     {
         class Test
         {
             public string InputString { get; set; }
             public bool ThrowsException { get; set; }
-            public UInt32 Result { get; set; }
+            public Int32 Result { get; set; }
 
-            public Test(string inputString, bool throwsException, UInt32 result)
+            public Test(string inputString, bool throwsException, Int32 result)
             {
                 InputString = inputString;
                 ThrowsException = throwsException;
                 Result = result;
             }
 
-            public Test(string inputString, UInt32 result)
+            public Test(string inputString, Int32 result)
             {
                 InputString = inputString;
                 ThrowsException = false;
@@ -35,18 +35,18 @@ namespace ConverterTestShared
         {
             new Test("0", 0),
             new Test("1", 1),
-            new Test("-1"),
+            new Test("-1", -1),
 
             new Test("255", Byte.MaxValue),
-            new Test("-128"),
-            new Test("127", 127),
+            new Test("-128", SByte.MinValue),
+            new Test("127", SByte.MaxValue),
 
             new Test("65535", UInt16.MaxValue),
-            new Test("-32768"),
-            new Test("32767", 32767),
+            new Test("-32768", Int16.MinValue),
+            new Test("32767", Int16.MaxValue),
 
-            new Test("4294967295", UInt32.MaxValue),
-            new Test("-2147483648"),
+            new Test("4294967295"),
+            new Test("-2147483648", Int32.MinValue),
             new Test("2147483647", Int32.MaxValue),
 
             new Test("18446744073709551615"),
@@ -85,7 +85,7 @@ namespace ConverterTestShared
 
                 try
                 {
-                    var val = UInt32.Parse(test.InputString);
+                    var val = Int32.Parse(test.InputString);
                     correctValue = (val == test.Result);
                 }
                 catch
@@ -107,7 +107,7 @@ namespace ConverterTestShared
                 }
 
             }
-            Console.WriteLine("TestUInt32 Tests: " + _testCount + " Fails: " + _fails);
+            Console.WriteLine("TestInt32 Tests: " + _testCount + " Fails: " + _fails);
         }
 
     }
